renaming package and module
